+++ zoo-modules/link-module/Link.svelte
@@ -13,6 +13,7 @@
 
 	:host {
 		contain: layout;
+		display: flex;
 	}
 
 	.box {